split up credentials.test.ts into multiple smaller test files with smaller setups
