+++ packages/cli/test/integration/credentials/generateUniqueName.test.ts
@@ -1,0 +1,70 @@
+import type { SuperAgentTest } from 'supertest';
+
+import config from '@/config';
+import type { User } from '@db/entities/User';
+import { License } from '@/License';
+
+import { randomCredentialPayload } from '../shared/random';
+import * as testDb from '../shared/testDb';
+import type { SaveCredentialFunction } from '../shared/types';
+import * as utils from '../shared/utils/';
+import { affixRoleToSaveCredential } from '../shared/db/credentials';
+import { createUser } from '../shared/db/users';
+
+// mock that credentialsSharing is not enabled
+jest.spyOn(License.prototype, 'isSharingEnabled').mockReturnValue(false);
+const testServer = utils.setupTestServer({ endpointGroups: ['credentials'] });
+
+let owner: User;
+let authOwnerAgent: SuperAgentTest;
+let saveCredential: SaveCredentialFunction;
+
+beforeAll(async () => {
+	owner = await createUser({ role: 'global:owner' });
+
+	saveCredential = affixRoleToSaveCredential('credential:owner');
+
+	authOwnerAgent = testServer.authAgentFor(owner);
+});
+
+beforeEach(async () => {
+	await testDb.truncate(['SharedCredentials', 'Credentials']);
+});
+
+describe('GET /credentials/new', () => {
+	test('should return default name for new credential or its increment', async () => {
+		const name = config.getEnv('credentials.defaultName');
+		let tempName = name;
+
+		for (let i = 0; i < 4; i++) {
+			const response = await authOwnerAgent.get(`/credentials/new?name=${name}`);
+
+			expect(response.statusCode).toBe(200);
+			if (i === 0) {
+				expect(response.body.data.name).toBe(name);
+			} else {
+				tempName = name + ' ' + (i + 1);
+				expect(response.body.data.name).toBe(tempName);
+			}
+			await saveCredential({ ...randomCredentialPayload(), name: tempName }, { user: owner });
+		}
+	});
+
+	test('should return name from query for new credential or its increment', async () => {
+		const name = 'special credential name';
+		let tempName = name;
+
+		for (let i = 0; i < 4; i++) {
+			const response = await authOwnerAgent.get(`/credentials/new?name=${name}`);
+
+			expect(response.statusCode).toBe(200);
+			if (i === 0) {
+				expect(response.body.data.name).toBe(name);
+			} else {
+				tempName = name + ' ' + (i + 1);
+				expect(response.body.data.name).toBe(tempName);
+			}
+			await saveCredential({ ...randomCredentialPayload(), name: tempName }, { user: owner });
+		}
+	});
+});
