+++ packages/cli/test/integration/credentials/getMany.test.ts
@@ -1,0 +1,81 @@
+import type { SuperAgentTest } from 'supertest';
+
+import type { ListQuery } from '@/requests';
+import type { User } from '@db/entities/User';
+import { License } from '@/License';
+
+import { randomCredentialPayload } from '../shared/random';
+import * as testDb from '../shared/testDb';
+import type { SaveCredentialFunction } from '../shared/types';
+import * as utils from '../shared/utils/';
+import { affixRoleToSaveCredential } from '../shared/db/credentials';
+import { createManyUsers, createUser } from '../shared/db/users';
+import { validateMainCredentialData } from './shared';
+
+// mock that credentialsSharing is not enabled
+jest.spyOn(License.prototype, 'isSharingEnabled').mockReturnValue(false);
+const testServer = utils.setupTestServer({ endpointGroups: ['credentials'] });
+
+let owner: User;
+let member: User;
+let authOwnerAgent: SuperAgentTest;
+let saveCredential: SaveCredentialFunction;
+
+beforeAll(async () => {
+	owner = await createUser({ role: 'global:owner' });
+	member = await createUser({ role: 'global:member' });
+
+	saveCredential = affixRoleToSaveCredential('credential:owner');
+
+	authOwnerAgent = testServer.authAgentFor(owner);
+});
+
+beforeEach(async () => {
+	await testDb.truncate(['SharedCredentials', 'Credentials']);
+});
+
+// ----------------------------------------
+// GET /credentials - fetch all credentials
+// ----------------------------------------
+describe('GET /credentials', () => {
+	test('should return all creds for owner', async () => {
+		const [{ id: savedOwnerCredentialId }, { id: savedMemberCredentialId }] = await Promise.all([
+			saveCredential(randomCredentialPayload(), { user: owner }),
+			saveCredential(randomCredentialPayload(), { user: member }),
+		]);
+
+		const response = await authOwnerAgent.get('/credentials');
+
+		expect(response.statusCode).toBe(200);
+		expect(response.body.data.length).toBe(2); // owner retrieved owner cred and member cred
+
+		const savedCredentialsIds = [savedOwnerCredentialId, savedMemberCredentialId];
+		response.body.data.forEach((credential: ListQuery.Credentials.WithOwnedByAndSharedWith) => {
+			validateMainCredentialData(credential);
+			expect('data' in credential).toBe(false);
+			expect(savedCredentialsIds).toContain(credential.id);
+		});
+	});
+
+	test('should return only own creds for member', async () => {
+		const [member1, member2] = await createManyUsers(2, {
+			role: 'global:member',
+		});
+
+		const [savedCredential1] = await Promise.all([
+			saveCredential(randomCredentialPayload(), { user: member1 }),
+			saveCredential(randomCredentialPayload(), { user: member2 }),
+		]);
+
+		const response = await testServer.authAgentFor(member1).get('/credentials');
+
+		expect(response.statusCode).toBe(200);
+		expect(response.body.data.length).toBe(1); // member retrieved only own cred
+
+		const [member1Credential] = response.body.data;
+
+		validateMainCredentialData(member1Credential);
+		expect(member1Credential.data).toBeUndefined();
+		expect(member1Credential.id).toBe(savedCredential1.id);
+	});
+});
